Split out GET tests
